# openapi.yaml
openapi: 3.0.3
info:
  title: RGB Lightning Node
  description: |-
    This is the OpenAPI specification for the
    [RGB Lightning Node](https://github.com/RGB-Tools/rgb-lightning-node) APIs.
  license:
    name: MIT
    url: https://mit-license.org/
  version: 0.1.0
servers:
  - url: http://localhost:3001
  - url: http://localhost:3002
  - url: http://localhost:3003
...
paths:
...
  /assetbalance:
    post:
      tags:
        - RGB
      summary: Get the balance of an asset
      description: Get the balance for the provided RGB asset
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetBalanceRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetBalanceResponse'
  /assetmetadata:
    post:
      tags:
        - RGB
      summary: Get the metadata of an asset
      description: Get the metadata for the provided RGB asset
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetMetadataRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetMetadataResponse'
  /backup:
    post:
      tags:
        - Other
      summary: Backup the node
      description: Create a backup of the node's data
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BackupRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
  /btcbalance:
    post:
      tags:
        - On-chain
      summary: Get the BTC balance
      description: Get the node's bitcoin balance for the vanilla and colored wallets
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BtcBalanceRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BtcBalanceResponse'
  /cancelhodlinvoice:
    post:
      tags:
        - Invoices
      summary: Cancel a HODL invoice
      description: Cancel a held HTLC for a HODL invoice. Rejects cancellation if a settlement is already in progress.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceCancelRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
  /changepassword:
    post:
      tags:
        - Other
      summary: Change the password
      description: Change the node's password
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
  /checkindexerurl:
    post:
      tags:
        - Other
      summary: Check an indexer URL
      description: Check the given indexer URL is valid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckIndexerUrlRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckIndexerUrlResponse'
  /checkproxyendpoint:
    post:
      tags:
        - Other
      summary: Check a proxy endpoint
      description: Check the given proxy endpoint is valid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckProxyEndpointRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
  /closechannel:
    post:
      tags:
        - Channels
      summary: Close a channel
      description: Close a LN channel cooperatively or forcibly
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CloseChannelRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
  /connectpeer:
    post:
      tags:
        - Peers
      summary: Connect to a peer
      description: Connect to the provided LN peer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConnectPeerRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
  /createutxos:
    post:
      tags:
        - RGB
      summary: Create UTXOs
      description: Create UTXOs to be used for RGB operations
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUtxosRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
  /decodelninvoice:
    post:
      tags:
        - Invoices
      summary: Decode a LN invoice
      description: Decode the provided LN invoice string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecodeLNInvoiceRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecodeLNInvoiceResponse'
  /decodergbinvoice:
    post:
      tags:
        - RGB
      summary: Decode an RGB invoice
      description: Decode the provided RGB invoice string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecodeRGBInvoiceRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecodeRGBInvoiceResponse'
  /disconnectpeer:
    post:
      tags:
        - Peers
      summary: Disconnect from a peer
      description: Disconnect from the provided LN peer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DisconnectPeerRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
  /estimatefee:
    post:
      tags:
        - On-chain
      summary: Get fee estimation
      description: Get on-chain fee estimation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EstimateFeeRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstimateFeeResponse'
  /failtransfers:
    post:
      tags:
        - RGB
      summary: Fail RGB transfers
      description: Set the status for eligible RGB transfers to
        `TransferStatus::Failed`.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FailTransfersRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FailTransfersResponse'
  /getassetmedia:
    post:
      tags:
        - RGB
      summary: Get an asset media
      description: Get the hex string of the media bytes of the provided media digest
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetAssetMediaRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAssetMediaResponse'
  /getchannelid:
    post:
      tags:
        - Channels
      summary: Get a channel's ID
      description: Get a channel's ID from its former temporary channel ID
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetChannelIdRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetChannelIdResponse'
  /getpayment:
    post:
      tags:
        - Payments
      summary: Get a payment
      description: Get a payment by its payment hash
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetPaymentRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetPaymentResponse'
  /getpaymentpreimage:
    post:
      tags:
        - Payments
      summary: Get a payment preimage by its payment hash
      description: Get the preimage for an outbound payment when it has been completed successfully
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetPaymentPreimageRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetPaymentPreimageResponse'
  /getswap:
    post:
      tags:
        - Swaps
      summary: Get a swap
      description: Get a swap by its payment hash
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetSwapRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetSwapResponse'
  /hodlinvoice:
    post:
      tags:
        - Invoices
      summary: Create a HODL LN invoice
      description: Create a BOLT11 invoice with a caller-provided payment hash; settlement is deferred until settle/cancel. Metadata is persisted first to preserve HODL semantics across restarts.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceHodlRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceHodlResponse'
  /init:
    post:
      tags:
        - Other
      summary: Init the node
      description: Initialize a new node
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InitRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InitResponse'
  /invoicestatus:
    post:
      tags:
        - Invoices
      summary: Get an invoice status
      description: Get the status of the provided LN invoice
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceStatusRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceStatusResponse'
  /issueassetcfa:
    post:
      tags:
        - RGB
      summary: Issue an RGB CFA asset
      description: Issue an RGB CFA asset. To provide a media first call the /postassetmedia API.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssueAssetCFARequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueAssetCFAResponse'
  /issueassetnia:
    post:
      tags:
        - RGB
      summary: Issue an RGB NIA asset
      description: Issue an RGB NIA asset
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssueAssetNIARequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueAssetNIAResponse'
  /issueassetuda:
    post:
      tags:
        - RGB
      summary: Issue an RGB UDA asset
      description: Issue an RGB UDA asset. To provide a media first call the /postassetmedia API.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssueAssetUDARequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueAssetUDAResponse'
  /keysend:
    post:
      tags:
        - Payments
      summary: Send to a peer spontaneously
      description: Send bitcoins and RGB assets to a LN peer spontaneously (without a LN invoice)
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KeysendRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KeysendResponse'
  /listassets:
    post:
      tags:
        - RGB
      summary: List assets
      description: List the node's RGB assets
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListAssetsRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListAssetsResponse'
...
  /listtransactions:
    post:
      tags:
        - On-chain
      summary: List transactions
      description: List the node's on-chain transactions
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListTransactionsRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListTransactionsResponse'
  /listtransfers:
    post:
      tags:
        - RGB
      summary: List transfers
      description: List the node's on-chain RGB transfers
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListTransfersRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListTransfersResponse'
  /listunspents:
    post:
      tags:
        - On-chain
      summary: List unspents
      description: List the unspent outputs of the internal BDK wallet
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListUnspentsRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListUnspentsResponse'
...
  /lninvoice:
    post:
      tags:
        - Invoices
      summary: Get a LN invoice
      description: Get a LN invoice to receive a payment
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LNInvoiceRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LNInvoiceResponse'
  /makerexecute:
    post:
      tags:
        - Swaps
      summary: Execute a maker swap
      description: Execute a swap on the maker side
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MakerExecuteRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
  /makerinit:
    post:
      tags:
        - Swaps
      summary: Init a maker swap
      description: Init a swap on the maker side
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MakerInitRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MakerInitResponse'
...
  /openchannel:
    post:
      tags:
        - Channels
      summary: Open a channel
      description: Open a new LN channel (RGB-enabled when both asset_id and asset_amount are specified).
        You can optionally provide a 32 bytes temporary channel ID as a hex-encoded string.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenChannelRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenChannelResponse'
  /postassetmedia:
    post:
      tags:
        - RGB
      summary: Post an asset media
      description: Save the provided media
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PostAssetMediaRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostAssetMediaResponse'
  /refreshtransfers:
    post:
      tags:
        - RGB
      summary: Refresh transfers
      description: Refresh RGB pending transfers
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
  /restore:
    post:
      tags:
        - Other
      summary: Restore the node
      description: Restore a node from a backup file
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestoreRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
  /revoketoken:
    post:
      tags:
        - Other
      summary: Revoke a token
      description: Revoke an authentication token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RevokeTokenRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
  /rgbinvoice:
    post:
      tags:
        - RGB
      summary: Get an RGB invoice
      description: Get an RGB invoice to receive assets on-chain
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RgbInvoiceRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RgbInvoiceResponse'
  /sendbtc:
    post:
      tags:
        - On-chain
      summary: Send BTC
      description: Send bitcoins on-chain
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendBtcRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendBtcResponse'
  /sendonionmessage:
    post:
      tags:
        - Other
      summary: Send an onion message
      description: Send an onion message via the LN
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendOnionMessageRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
  /sendpayment:
    post:
      tags:
        - Payments
      summary: Send a payment
      description: Pay the provided LN invoice
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendPaymentRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendPaymentResponse'
  /sendrgb:
    post:
      tags:
        - RGB
      summary: Send RGB assets
      description: Send RGB assets on-chain, supporting batch transfers to multiple recipients and/or multiple assets in a single transaction
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendRgbRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendRgbResponse'
  /settlehodlinvoice:
    post:
      tags:
        - Invoices
      summary: Settle a HODL invoice
      description: Claim a held HTLC for a HODL invoice
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceSettleRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
...
  /signmessage:
    post:
      tags:
        - Other
      summary: Sign a message
      description: Sign the provided message
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignMessageRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignMessageResponse'
...
  /taker:
    post:
      tags:
        - Swaps
      summary: Accept a swap
      description: Accept a swap on the taker side
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TakerRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
...
components:
  schemas:
...
    AssetBalanceRequest:
      type: object
      properties:
        asset_id:
          type: string
          example: rgb:CJkb4YZw-jRiz2sk-~PARPio-wtVYI1c-XAEYCqO-wTfvRZ8
    AssetBalanceResponse:
      type: object
      properties:
        settled:
          type: integer
          example: 777
        future:
          type: integer
          example: 777
        spendable:
          type: integer
          example: 777
        offchain_outbound:
          type: integer
          example: 444
        offchain_inbound:
          type: integer
          example: 0
...
    AssetMetadataRequest:
      type: object
      properties:
        asset_id:
          type: string
          example: rgb:CJkb4YZw-jRiz2sk-~PARPio-wtVYI1c-XAEYCqO-wTfvRZ8
...
    BackupRequest:
      type: object
      properties:
        backup_path:
          type: string
          example: /path/where/to/save/the/backup/file
        password:
          type: string
          example: nodepassword
...
    BtcBalanceRequest:
      type: object
      properties:
        skip_sync:
          type: boolean
          example: false
    BtcBalanceResponse:
      type: object
      properties:
        vanilla:
          $ref: '#/components/schemas/BtcBalance'
        colored:
          $ref: '#/components/schemas/BtcBalance'
...
    EmptyResponse:
      type: object
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Biscuit
security:
  - bearerAuth: []